Add SlowCallback error type
TestSlide previously raised a RuntimeError when it detected an asyncio
slow callback warning message, change that to a custom SlowCallback
exception type that makes the failure case easier to track down, since
the default message is fairly opaque if you haven't seen it before.

This could potentially be extended with more information to help track
down the offending function, but Python does not provide much
information, so in this PR I am just adding the exception class to at
least point in the right direction.