[signature validation] Skip TypeVar signatures
In order to be able to check TypeVars we have to be able to check the
whole signature type hints at the same time (as typevars are dynamically
defined).

This just skips them for now to avoid failing the type checks.

It's using the `str` method of the type hint to traverse the type
definition and see if there are any typevars (prefixed with `~`), it's a
hacky way, but quite effective until we have the need to traverse the
type hint ourselves, at that point we can use that instead.