[global] add isort and `make format`
We were only checking black, this adds also isort with black compatible
settings and adds a new make subcommand `format` to apply all the
changes to all the files (we don't have too many yet to do it only on
the changed files only).

Also moved all the relative import to non-relative when possible (just
standardizing on something).

This pins isort to version >5, that means that the support for python
3.5 is being dropped (it's not in the list of tested environments).
See:
https://timothycrosley.github.io/isort/docs/major_releases/introducing_isort_5/

It also make some of the cli_unittest lest flaky by using regexes to
parse output instead of hardcoded line numbers.

Signed-off-by: David Caro <david@dcaro.es>