Raise Exception from AnyInstanceOf Constructor (#291)
Summary:
<!--
Thank you for your interest in this project. Bugs filed and PRs submitted are appreciated!

Please make sure that you are familiar with and follow the Code of Conduct of this project which can be found at https://github.com/facebookincubator/TestSlide/blob/master/CODE_OF_CONDUCT.md

Also, please make sure you're familiar with and follow the instructions in the contributing guidelines which can  be found at https://github.com/facebookincubator/TestSlide/blob/master/CONTRIBUTING.md

-->

**What:**

https://github.com/facebook/TestSlide/issues/233

**Why:**

**How:**

Checked for the input arguments of the `AnyInstanceOf` to be of an instance of `type`. If not, raise a `ValueError` exception.

**Risks:**

**Checklist**:

<!--
Have you done all of these things?
To check an item, place an "x" in the box like so: "- [x] Tests"
Add "N/A" to the end of each line that's irrelevant to your changes
-->

- [x] Added tests, if you've added code that should be tested
- [x] Updated the documentation, if you've changed APIs
- [x] Ensured the test suite passes
- [x] Made sure your code lints
- [x] Completed the Contributor License Agreement ("CLA")

Do I need to add the "correct usage" of the AnyInstanceOf in the tests?

Pull Request resolved: https://github.com/facebook/TestSlide/pull/291

Reviewed By: deathowl

Differential Revision: D27237895

Pulled By: macisamuele

fbshipit-source-id: a13a14d0cd04e747a143f2c6722b285e0cf17625