Ensure that type validation is executed for StrictMock instances as well (#285)
Summary:
<!--
Thank you for your interest in this project. Bugs filed and PRs submitted are appreciated!

Please make sure that you are familiar with and follow the Code of Conduct of this project which can be found at https://github.com/facebookincubator/TestSlide/blob/master/CODE_OF_CONDUCT.md

Also, please make sure you're familiar with and follow the instructions in the contributing guidelines which can  be found at https://github.com/facebookincubator/TestSlide/blob/master/CONTRIBUTING.md

-->

While debugging some issue into some FB internal tests (which I was expecting failing) I've noticed that using `mock_callable` / `mock_async_callable` on `StrictMock` does not enable type validation.
This feels like an oversight on the code and has the potential to have users (myself included) writing tests expecting the framework to be validating the types (and as such ensuring that the mocks are consistent with the typing annotations).

A super small representative example is available on https://gist.github.com/macisamuele/b6e85ddfad7f5802ebedd5a2a0810c6c

This PR does also fixes/improves few more details:
* ensure that functions wrapped for type checking are not altering function name and docstring (debuggability)
* removes redundant `FORCE` Makefile target
* updates the repository link in `setup.py`

## What:

## Why:

These changes are needed to ensure that `mock_callable` and `mock_async_callable` does preserve type validation regardless of the target being a *real* instance or a `StrictMock` instance.

**NOTE**: In order to make the change possible and well defined it does modify the `tests.sample_module` (6be4622482e77b45ce240ffbf8804ac75fbb30ae) in order to have the methods returning `Iterable`s instead of `str`s.
This is needed because current testing (`mock_callable_testslide.py`) does mock `class_method` with `to_yield_values` which is incompatible from a typing prospective. Fixing the tests would have required more work respect fixing the module class and it is safe because the module is used only for testing purposes.

## How:

The `mock_callable` implementation was intentionally excluding `StrictMock` instances to be wrapped by type validation, this PR *just* removed this exclusion.

## Risks:

Users of the library which have been relying on the lack of type validation on mocked functions on `StrictMock`s would see their tests failing.
This is intentional as the lack of the type validation in such cases was a bug and not considered a feature.
For these users, my suggestion would be to fix their tests or to add `type_validation=False` into the `mock_(async_)callable` definition (the later is discouraged as it hides the issue instead of solving it)

## Checklist

<!--
Have you done all of these things?
To check an item, place an "x" in the box like so: "- [x] Tests"
Add "N/A" to the end of each line that's irrelevant to your changes
-->

- [x] Added tests, if you've added code that should be tested
- [ ] Updated the documentation, if you've changed APIs
- [x] Ensured the test suite passes
- [x] Made sure your code lints
- [x] Completed the Contributor License Agreement ("CLA")

Pull Request resolved: https://github.com/facebook/TestSlide/pull/285

Reviewed By: deathowl

Differential Revision: D26568686

Pulled By: macisamuele

fbshipit-source-id: 56e329823852a536c7638f63bbf5e6b8dd07a0a2