Enable running all tests to completion even if one python version failed (#295)
Summary:
**What:**
Ensure that we do run all the tests (platforms/python-versions) and that we don't immediately stop if one has failed.
This is a follow-up from https://github.com/facebook/TestSlide/issues/292

**Why:**
In case of test failures is extremely interesting to have information about "is this failure a genuine bug or is an issue with a single python version?".

**How:**
Ensure that github workflow is correctly defined. Doc in [here](https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#jobsjob_idstrategyfail-fast)

**Risks:**

n/a

**Checklist**:

- [x] Added tests, if you've added code that should be tested
- [x] Updated the documentation, if you've changed APIs
- [x] Ensured the test suite passes
- [x] Made sure your code lints
- [x] Completed the Contributor License Agreement ("CLA")

Pull Request resolved: https://github.com/facebook/TestSlide/pull/295

Reviewed By: deathowl

Differential Revision: D27395399

Pulled By: macisamuele

fbshipit-source-id: e474f46614b5422c798b0fcfbd73edab58a9e007