Fix incorrect hiding for stack frames when reporting exceptions
Summary:
Tests can fail on setup/tearDown or with exceptions that were originally thrown in an asyncio task and then used as `__cause__`

Testslide's code has a bug, where if the test's module does not appear in the traceback, only the topmost frame is printed.

This diff fixes the bug.

NOTE: actually, even this diff is a half-measure, it is possible e.g. using callbacks to create stack traces that begin in the code under test, and only then reach back into the test module. The code will still erase such frames, which is very bad. Perhaps the better diff is to get rid of this heuristic entirely.

Differential Revision: D28771793

fbshipit-source-id: 4d6cd6012e9ee0218f230512d17a1dbf8e3fe570