add py310-dev to the build matrix (#310)
Summary:
**What:**
Add 3.10-dev to the build matrix.
**Why:**
Let's start getting builds for 3.10 early
**How:**

Added 3.10-dev to the github actions file
**Risks:**

None

**Checklist**:

<!--
Have you done all of these things?
To check an item, place an "x" in the box like so: "- [x] Tests"
Add "N/A" to the end of each line that's irrelevant to your changes
-->

- [X] Added tests, if you've added code that should be tested
- [X] Updated the documentation, if you've changed APIs
- [X] Ensured the test suite passes
- [X] Made sure your code lints
- [X] Completed the Contributor License Agreement ("CLA")

Pull Request resolved: https://github.com/facebook/TestSlide/pull/310

Reviewed By: kawmarco

Differential Revision: D29687577

Pulled By: deathowl

fbshipit-source-id: 1619bfc3466f26fa84e66b49727b96f982bd1100