Avoid fataling on StrictMock(template=str).garbage
Summary:
Accessing nonexistent attributes of `StrictMock`s templated on `str` causes fatals.
This is because of an unfortunate self-reference, where when a `UndefinedAttribute` or `AttributeError` is thrown, it tries to include `f"{self}"`, which invokes `self.__format__`, which hits the attribute handling, which triggers another UndefinedAttribute...

The solution is just to directly point those prints to `repr(self`)`, which works.

Reviewed By: deathowl

Differential Revision: D30966913

fbshipit-source-id: fd84a5ac256b9010acb676a32c6b9b4f8bba7fe0