Add another backtrack to stack frame in mocks
Summary:
**Problem**

The testslide mocking functions are usually called not directly
via the `testslide`. As a result, when I get errors such as a return type
mismatch, it's taking me to `testslide.__init__` as the source of the problem,
which is not at all useful. I want it to take me to the actual test where the bad
mock was defined, and to do that I need to go one more stack frame up.

**Simple solution (the current diff)**

Assume that mocking is *always* done through the TestCase
methods. In that case, we can just backtrack 2 frames instead of one to
find the definition.

**An alternative solution if we cannot break backward compatibility to direct calls**

I don't know the testslide API terribly well, so I don't have
a good sense whether callers would directly use testslide.mock_callable.mock_callable.

If that's the case, then I need to make this change backward compatible; happy
to do that, it would require adding an extra argument indicating how far back
to go with a default value of 1, and we'd override to 2 in all of the methods of
TestCase.

Since that's a bit more complicated, I'm going to wait for diff comments before I do that.

Reviewed By: deathowl

Differential Revision: D33870723

fbshipit-source-id: bf17f385f82e8df5e0aaccf75b1f804d3463e50e