Fix default_context_manager in conjunction with ExitStack. (#344)
Summary:
Pull Request resolved: https://github.com/facebook/TestSlide/pull/344

As outlined in the comment; because our MethodProxy objects don't get bound to
the object (via __get__); when (Async)ExitStack comes along and invokes:

```
enter_fn = cm.__class_.__enter__
enter_fn(cm)
```

A problem occurs - while the attribute lookup succeeds (yay); it's not
expecting a `self` paramter explicitly passed here, as it's not a bound python
method.

Theoretically; magic could be added to MethodProxy to properly bind via
`__get__`, but my knowledge of testslide right now isn't strong enough to feel
confident in that approach, so I'm opting for this.

Reviewed By: deathowl

Differential Revision: D42114206

fbshipit-source-id: 9473fe59d30c93c56717f92117180f1b6f9e31c1