Remove circular testslide imports
Summary:
While testing the import cost for some binaries I've found out that testslide is providing an implicit circular import.
An example is `testslide.mock_constructor` importing `testslide`.

This is not necessarily an issue, but I would love to improve it.

In the general case the `testslide` import was needed:
* to make `testslide._importer` available, I'm fixing it via a local import
* to have `testslide.__file__`, even in this case a local import would be sufficient

Differential Revision: D51025476

fbshipit-source-id: b3292393137c01a6a64b29d620dadb001a989260