Fix mocking on unhashable objs
Summary:
Ran into a nuanced bug that was triggered by multiple
`self.mock_callable` calls for the same function on an unhashable object
resulting in
```
RuntimeError: Exceptions raised when unpatching: [TypeError("unhashable type:
'SomeUnhashableClass'")]
```

The issue was caused by the interaction of two bugs:

1) When checking whether a method is an instance method, the `_DescriptorProxy`
value isn't always properly unwrapped. This unwrapping is necessary because
multiple mocking calls to a given attribute result in nested
`_DescriptorProxies`. When this doesn't happen, we can wrongly return False
from `_is_instance_method`.

2) `_DescriptorProxy` deletions do not work on unhashable types due to the
check we do for the instance within `instance_attr_map`. By calling
`id(instance)` we'll be able to support these unhashable objects and be more
consistent with the rest of the code that does similar checks.

Reviewed By: deathowl

Differential Revision: D54349054

fbshipit-source-id: 80fe70f93cd44fc5458bc5229d27dc967b4be83a